refactor: style 파일 styles로 분리
